# perf(): read shards in parallel for Iter, IterBuffered and Keys