This PR spawns a goroutine for each shard when iterating them. This increases performance in two ways:

1) With multiple cores, items can be read from shards in parallel (specifically with the buffered iterator)
2) If one of the shards is locked it could block all subsequent reads. This allows all the other shards to be read, even if one of them is locked.


This may also reduce the likelihood of running into some of the race conditions mentioned in #6 since all of will be locked at close to the same time when the goroutines spawn.

Informal benchmark results are as follows:

**Before**

```
PASS
BenchmarkItems-4                                     300           4536721 ns/op
BenchmarkMarshalJson-4                               100          18127612 ns/op
BenchmarkStrconv-4                              20000000                66.8 ns/op
BenchmarkSingleInsertAbsent-4                    1000000              1131 ns/op
BenchmarkSingleInsertPresent-4                  10000000               230 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4           500000              3921 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4          500000              3898 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4          500000              3938 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4         300000              5420 ns/op
BenchmarkMultiInsertSame-4                        500000              2434 ns/op
BenchmarkMultiGetSame-4                          1000000              1421 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4           300000              5544 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4          300000              5430 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4          300000              5379 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4         300000              5513 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4               300000              4284 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4              300000              4083 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4              300000              4036 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4             300000              3929 ns/op
BenchmarkKeys-4                                     1000           1687293 ns/op
ok      github.com/streamrail/concurrent-map    35.428s
```

**After**
```
PASS
BenchmarkItems-4                                     300           5739354 ns/op
BenchmarkMarshalJson-4                               100          19056413 ns/op
BenchmarkStrconv-4                              20000000                67.4 ns/op
BenchmarkSingleInsertAbsent-4                    1000000              1218 ns/op
BenchmarkSingleInsertPresent-4                   5000000               227 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4           500000              3920 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4          300000              3840 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4          500000              3921 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4         200000              5673 ns/op
BenchmarkMultiInsertSame-4                        500000              2388 ns/op
BenchmarkMultiGetSame-4                          1000000              1367 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4           300000              5476 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4          300000              5475 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4          300000              5318 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4         300000              5391 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4               300000              4236 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4              300000              4162 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4              300000              4015 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4             300000              3937 ns/op
BenchmarkKeys-4                                      500           2332856 ns/op
ok      github.com/streamrail/concurrent-map    32.107s
```