@redbaron 2016-08-13
None of tests show improvement and BenchmarkKeys shows 1.5x slowdown, was it worth it?